refactor(app): useLPCDisabledReason from pages
