+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/ActionButton/index.tsx
@@ -14,10 +14,7 @@ import {
   useHoverTooltip,
 } from '@opentrons/components'
 
-import {
-  useModuleCalibrationStatus,
-  useUnmatchedModulesForProtocol,
-} from '../../../../hooks'
+import { useModuleCalibrationStatus } from '../../../../hooks'
 import { useRobot } from '/app/redux-resources/robots'
 import { useRobotAnalyticsData } from '/app/redux-resources/analytics'
 import {
@@ -25,6 +22,7 @@ import {
   useCurrentRunId,
   useProtocolDetailsForRun,
   useRunCalibrationStatus,
+  useUnmatchedModulesForProtocol,
 } from '/app/resources/runs'
 import { useActionBtnDisabledUtils, useActionButtonProperties } from './hooks'
 import { getFallbackRobotSerialNumber, isRunAgainStatus } from '../../utils'
