+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunModuleControls.tsx
@@ -9,9 +9,9 @@ import {
   InfoScreen,
   SPACING,
 } from '@opentrons/components'
-import { ModuleCard } from '../../ModuleCard'
-import { useModuleRenderInfoForProtocolById } from '../hooks'
-import { useModuleApiRequests } from '../../ModuleCard/utils'
+import { ModuleCard } from '/app/organisms/ModuleCard'
+import { useModuleRenderInfoForProtocolById } from '/app/resources/runs'
+import { useModuleApiRequests } from '/app/organisms/ModuleCard/utils'
 
 import type { BadPipette, PipetteData } from '@opentrons/api-client'
