+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunSetup.tsx
@@ -31,16 +31,13 @@ import {
 import { useDeckConfigurationCompatibility } from '/app/resources/deck_configuration/hooks'
 import { useRobot, useIsFlex } from '/app/redux-resources/robots'
 import { useStoredProtocolAnalysis } from '/app/resources/analysis'
-import {
-  useModuleCalibrationStatus,
-  useProtocolAnalysisErrors,
-  useUnmatchedModulesForProtocol,
-} from '../hooks'
+import { useModuleCalibrationStatus, useProtocolAnalysisErrors } from '../hooks'
 import {
   useMostRecentCompletedAnalysis,
   useRunPipetteInfoByMount,
   useRunCalibrationStatus,
   useRunHasStarted,
+  useUnmatchedModulesForProtocol,
 } from '/app/resources/runs'
 import { SetupLabware } from './SetupLabware'
 import { SetupLabwarePositionCheck } from './SetupLabwarePositionCheck'
