+++ app/src/organisms/Devices/ProtocolRun/SetupLabwarePositionCheck/__tests__/CurrentOffsetsTable.test.tsx
@@ -11,14 +11,14 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { getIsLabwareOffsetCodeSnippetsOn } from '/app/redux/config'
 import { LabwarePositionCheck } from '../../../../LabwarePositionCheck'
-import { useLPCDisabledReason } from '../../../hooks'
+import { useLPCDisabledReason } from '/app/resources/runs'
 import { getLatestCurrentOffsets } from '/app/transformations/runs'
 import { CurrentOffsetsTable } from '../CurrentOffsetsTable'
 
 import type { CompletedProtocolAnalysis } from '@opentrons/shared-data'
 import type { LabwareOffset } from '@opentrons/api-client'
 
-vi.mock('../../../hooks')
+vi.mock('/app/resources/runs')
 vi.mock('../../../../LabwarePositionCheck')
 vi.mock('/app/redux/config')
 vi.mock('/app/transformations/runs')
