+++ app/src/organisms/Devices/ProtocolRun/SetupLabwarePositionCheck/__tests__/SetupLabwarePositionCheck.test.tsx
@@ -16,15 +16,13 @@ import { useLPCSuccessToast } from '../../../hooks/useLPCSuccessToast'
 import { getModuleTypesThatRequireExtraAttention } from '../../utils/getModuleTypesThatRequireExtraAttention'
 import { useLaunchLPC } from '../../../../LabwarePositionCheck/useLaunchLPC'
 import { getIsLabwareOffsetCodeSnippetsOn } from '/app/redux/config'
-import {
-  useLPCDisabledReason,
-  useUnmatchedModulesForProtocol,
-} from '../../../hooks'
 import { SetupLabwarePositionCheck } from '..'
 import {
   useNotifyRunQuery,
   useRunCalibrationStatus,
   useRunHasStarted,
+  useLPCDisabledReason,
+  useUnmatchedModulesForProtocol,
 } from '/app/resources/runs'
 import { useRobotType } from '/app/redux-resources/robots'
 
@@ -34,7 +32,6 @@ vi.mock('../../../../LabwarePositionCheck/useLaunchLPC')
 vi.mock('../../utils/getModuleTypesThatRequireExtraAttention')
 vi.mock('/app/redux-resources/robots')
 vi.mock('/app/redux/config')
-vi.mock('../../../hooks')
 vi.mock('../../../hooks/useLPCSuccessToast')
 vi.mock('@opentrons/react-api-client')
 vi.mock('/app/resources/runs')
