+++ app/src/organisms/Devices/ProtocolRun/__tests__/ProtocolRunModuleControls.test.tsx
@@ -7,8 +7,8 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { useInstrumentsQuery } from '@opentrons/react-api-client'
 import { ProtocolRunModuleControls } from '../ProtocolRunModuleControls'
-import { ModuleCard } from '../../../ModuleCard'
-import { useModuleRenderInfoForProtocolById } from '../../hooks'
+import { ModuleCard } from '/app/organisms/ModuleCard'
+import { useModuleRenderInfoForProtocolById } from '/app/resources/runs'
 import {
   mockMagneticModuleGen2,
   mockTemperatureModuleGen2,
@@ -18,8 +18,8 @@ import {
 import type { ModuleModel, ModuleType } from '@opentrons/shared-data'
 
 vi.mock('@opentrons/react-api-client')
-vi.mock('../../../ModuleCard')
-vi.mock('../../hooks')
+vi.mock('/app/organisms/ModuleCard')
+vi.mock('/app/resources/runs')
 
 const RUN_ID = 'test123'
 const mockTempMod = {
