+++ app/src/organisms/Devices/hooks/index.ts
@@ -5,14 +5,11 @@ export * from './useIsRobotBusy'
 export * from './useIsRobotViewable'
 export * from './useLights'
 export * from './useLEDLights'
-export * from './useLPCDisabledReason'
 export * from './useLPCSuccessToast'
-export * from './useModuleRenderInfoForProtocolById'
 export * from './useModuleCalibrationStatus'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
 export * from './useTipLengthCalibrations'
-export * from './useUnmatchedModulesForProtocol'
 export * from './useProtocolAnalysisErrors'
 export * from './useRunStartedOrLegacySessionInProgress'
 export * from './useTrackCreateProtocolRunEvent'
