+++ app/src/pages/Desktop/Devices/ProtocolRunDetails/__tests__/ProtocolRunDetails.test.tsx
@@ -7,11 +7,7 @@ import { when } from 'vitest-when'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
-import {
-  useModuleRenderInfoForProtocolById,
-  useRunStatuses,
-  useSyncRobotClock,
-} from '/app/organisms/Devices/hooks'
+import { useRunStatuses, useSyncRobotClock } from '/app/organisms/Devices/hooks'
 import { ProtocolRunHeader } from '/app/organisms/Devices/ProtocolRun/ProtocolRunHeader'
 import { ProtocolRunModuleControls } from '/app/organisms/Devices/ProtocolRun/ProtocolRunModuleControls'
 import { ProtocolRunSetup } from '/app/organisms/Devices/ProtocolRun/ProtocolRunSetup'
@@ -21,6 +17,7 @@ import {
   useCurrentRunId,
   useMostRecentCompletedAnalysis,
   useRunHasStarted,
+  useModuleRenderInfoForProtocolById,
 } from '/app/resources/runs'
 import { mockRobotSideAnalysis } from '/app/molecules/Command/__fixtures__'
 import { useRobot } from '/app/redux-resources/robots'
