+++ app/src/pages/ODD/ProtocolSetup/index.tsx
@@ -34,7 +34,6 @@ import {
 } from '@opentrons/shared-data'
 
 import {
-  useLPCDisabledReason,
   useModuleCalibrationStatus,
   useProtocolAnalysisErrors,
 } from '/app/organisms/Devices/hooks'
@@ -79,7 +78,11 @@ import { getLatestCurrentOffsets } from '/app/transformations/runs'
 import { CloseButton, PlayButton } from './Buttons'
 import { useDeckConfigurationCompatibility } from '/app/resources/deck_configuration/hooks'
 import { getRequiredDeckConfig } from '/app/resources/deck_configuration/utils'
-import { useNotifyRunQuery, useRunStatus } from '/app/resources/runs'
+import {
+  useNotifyRunQuery,
+  useRunStatus,
+  useLPCDisabledReason,
+} from '/app/resources/runs'
 
 import type { Run } from '@opentrons/api-client'
 import type { CutoutFixtureId, CutoutId } from '@opentrons/shared-data'
