+++ app/src/resources/runs/index.ts
@@ -22,3 +22,6 @@ export * from './useRunPipetteInfoByMount'
 export * from './useRunCalibrationStatus'
 export * from './useRunHasStarted'
 export * from './useRunCreatedAtTimestamp'
+export * from './useLPCDisabledReason'
+export * from './useUnmatchedModulesForProtocol'
+export * from './useModuleRenderInfoForProtocolById'
