+++ app/src/transformations/runs/onDeviceDisplayFormatTimestamp.ts
@@ -1,7 +1,7 @@
 import { format, parseISO } from 'date-fns'
 export function onDeviceDisplayFormatTimestamp(timestamp: string): string {
-    // eslint-disable-next-line eqeqeq
-    return (parseISO(timestamp) as Date | string) != 'Invalid Date'
-        ? format(parseISO(timestamp), 'HH:mm:ss')
-        : timestamp
+  // eslint-disable-next-line eqeqeq
+  return (parseISO(timestamp) as Date | string) != 'Invalid Date'
+    ? format(parseISO(timestamp), 'HH:mm:ss')
+    : timestamp
 }
